api/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.68.1
h11==0.12.0
kombu==5.1.0
pika==1.2.0
pika-stubs==0.1.3
prompt-toolkit==3.0.20
pydantic==1.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
